# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IPPEL RNC - API de Gestão de Não Conformidades
  description: |
    API RESTful completa para gestão de Relatórios de Não Conformidade (RNC) do sistema IPPEL.
    
    ## Funcionalidades Principais
    - 🔐 Autenticação JWT com 2FA opcional
    - 📋 CRUD completo de RNCs
    - 👥 Gerenciamento de usuários e grupos
    - 🔒 Sistema de permissões granulares (Field Locks)
    - 📊 Analytics e dashboards em tempo real
    - 💬 Sistema de chat e notificações
    - 📄 Geração de relatórios (PDF, Excel, CSV)
    
    ## Autenticação
    A maioria dos endpoints requer autenticação via JWT token no header:
    ```
    Authorization: Bearer {token}
    ```
    
    ## Rate Limiting
    - **Login**: 10 req/min por IP
    - **API Geral**: 120 req/min por usuário
    - **Relatórios**: 30 req/min por usuário
    
    ## Versionamento
    Esta é a **versão 1.0** da API. Futuras versões serão prefixadas com `/v2/`, `/v3/`, etc.
    
  version: 1.0.0
  contact:
    name: Equipe IPPEL
    email: suporte@ippel.com.br
  license:
    name: Proprietário
servers:
  - url: http://localhost:5001
    description: Desenvolvimento local
  - url: https://rnc.ippel.com.br
    description: Produção
...
security:
  - BearerAuth: []
paths:
...
  /api/rnc/{id}:
    get:
      tags: [RNC]
      summary: Visualizar RNC
      description: Retorna detalhes completos de uma RNC específica
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalhes da RNC
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  rnc:
                    $ref: '#/components/schemas/RNC'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [RNC]
      summary: Editar RNC
      description: Atualiza dados de uma RNC existente (respeitando field locks)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RNCUpdate'
      responses:
        '200':
          description: RNC atualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags: [RNC]
      summary: Deletar RNC
      description: Remove uma RNC (apenas admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: RNC deletada
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ==================== ADMIN ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RNC:
      type: object
      properties:
        id:
          type: integer
        rnc_number:
          type: string
          example: RNC-2025-001
        title:
          type: string
        description:
          type: string
        equipment:
          type: string
        client:
          type: string
        priority:
          type: string
          enum: [Baixa, Média, Alta, Crítica]
        status:
          type: string
          enum: [Pendente, Em Andamento, Finalizado]
        department:
          type: string
        created_by:
          type: integer
        created_at:
          type: string
          format: date-time
        finalized_at:
          type: string
          format: date-time
        disposition_usar:
          type: boolean
        disposition_retrabalhar:
          type: boolean
        disposition_rejeitar:
          type: boolean
        disposition_sucata:
          type: boolean
        inspection_approved:
          type: boolean
        inspection_rejected:
          type: boolean
        inspection_date:
          type: string
          format: date
        inspection_name:
          type: string
        engineering_date:
          type: string
          format: date
        engineering_name:
          type: string
        value:
          type: number
          format: float
...
    RNCUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Pendente, Em Andamento, Finalizado]
        priority:
          type: string
          enum: [Baixa, Média, Alta, Crítica]
        disposition_usar:
          type: boolean
        inspection_approved:
          type: boolean
        inspection_name:
          type: string
        engineering_name:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Erro descritivo
        errors:
          type: array
          items:
            type: string
  responses:
...
    ForbiddenError:
      description: Sem permissão para esta operação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Você não tem permissão para esta ação
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: RNC não encontrada
...